Split file handleChange to namespaces and move constructing api URL to service layer #JC19-352
