+++ apps/vanilla/src/services/fetchAnime.ts
@@ -8,6 +8,8 @@ import { paginationMapper } from '@js-camp/core/mappers/pagination.mapper';
 import { AnimeMapper } from '@js-camp/core/mappers/anime.mapper';
 
 import { http } from '../api';
+import { NEXT_PAGE_LS, PREVIOUS_PAGE_LS, COUNT_LS } from '../script/constants';
+import { renderAnime, renderPagination } from '../script/renderToDOM';
 
 /**
  * Fetch anime data from api.
@@ -21,4 +23,19 @@ export async function fetchAnime(url: string): Promise<Pagination<Anime>> {
   } catch (error: unknown) {
     throw Error((error as Error).message);
   }
+}
+
+/**
+ * Update Anime data after change something like sort or page.
+ * @param url Api address.
+ */
+export async function updateAnime(url: string): Promise<void> {
+  const data = await fetchAnime(url);
+
+  localStorage.setItem(NEXT_PAGE_LS, data.next ?? '');
+  localStorage.setItem(PREVIOUS_PAGE_LS, data.previous ?? '');
+  localStorage.setItem(COUNT_LS, data.count.toString());
+  renderAnime(data);
+
+  renderPagination();
 }
